ACPI: Convert ACPI reference args to generic fwnode reference args (fate#326250).
suse-commit: a48b9b8400d927557245c58dc7af9f899bc3850d

+++ include/linux/fwnode.h
@@ -44,7 +44,7 @@ struct fwnode_endpoint {
 struct fwnode_reference_args {
 	struct fwnode_handle *fwnode;
 	unsigned int nargs;
-	unsigned int args[NR_FWNODE_REFERENCE_ARGS];
+	u64 args[NR_FWNODE_REFERENCE_ARGS];
 };
 
 /**
